refactoring auth handling in utility class
