+++ core/src/main/kotlin/org/evomaster/core/remote/AuthenticationRequiredException.kt
@@ -1,0 +1,3 @@
+package org.evomaster.core.remote
+
+class AuthenticationRequiredException(message: String) : SutProblemException(message)
