+++ core/src/main/kotlin/org/evomaster/core/remote/SutProblemException.kt
@@ -1,6 +1,6 @@
 package org.evomaster.core.remote
 
 
-class SutProblemException(
+open class SutProblemException(
         message : String = "Problem with the system under test."
 ) : RuntimeException(message)
